+++ ErsatzTV.Core/Metadata/Nfo/MusicVideoNfo.cs
@@ -17,6 +17,9 @@ public class MusicVideoNfo
     [XmlElement("plot")]
     public string Plot { get; set; }
 
+    [XmlElement("aired")]
+    public Option<DateTime> Aired { get; set; }
+
     [XmlElement("year")]
     public int Year { get; set; }