+++ ErsatzTV.Infrastructure/Search/CustomQueryParser.cs
@@ -36,7 +36,7 @@ public class CustomQueryParser : QueryParser
         if (field == "released_onthisday")
         {
             var todayString = DateTime.Today.ToString("*MMdd");
-            return base.GetWildcardQuery("release_date", todayString);
+            return base.GetWildcardQuery(SearchIndex.ReleaseDateField, todayString);
         }
 
         if (NumericFields.Contains(field) && int.TryParse(queryText, out int val))
@@ -56,14 +56,14 @@ public class CustomQueryParser : QueryParser
             var todayString = DateTime.UtcNow.ToString("yyyyMMdd");
             var dateString = start.ToString("yyyyMMdd");
 
-            return base.GetRangeQuery("release_date", dateString, todayString, true, true);
+            return base.GetRangeQuery(SearchIndex.ReleaseDateField, dateString, todayString, true, true);
         }
 
         if (field == "released_notinthelast" && ParseStart(queryText, out DateTime finish))
         {
             var dateString = finish.ToString("yyyyMMdd");
 
-            return base.GetRangeQuery("release_date", "00000000", dateString, false, false);
+            return base.GetRangeQuery(SearchIndex.ReleaseDateField, "00000000", dateString, false, false);
         }
 
         if (field == "added_inthelast" && ParseStart(queryText, out DateTime addedStart))
@@ -71,14 +71,14 @@ public class CustomQueryParser : QueryParser
             var todayString = DateTime.UtcNow.ToString("yyyyMMdd");
             var dateString = addedStart.ToString("yyyyMMdd");
 
-            return base.GetRangeQuery("added_date", dateString, todayString, true, true);
+            return base.GetRangeQuery(SearchIndex.AddedDateField, dateString, todayString, true, true);
         }
 
         if (field == "added_notinthelast" && ParseStart(queryText, out DateTime addedFinish))
         {
             var dateString = addedFinish.ToString("yyyyMMdd");
 
-            return base.GetRangeQuery("added_date", "00000000", dateString, false, false);
+            return base.GetRangeQuery(SearchIndex.AddedDateField, "00000000", dateString, false, false);
         }
 
         return base.GetFieldQuery(field, queryText, slop);